requirements.txt
langchain
openai
ccxt
tronpy
python-dotenv
pandas
